requirements.txt
PyYAML>=3.12
pymongo>=3.7.1
bson>=0.5.8